pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "visual-deep-learning"
version = "0.1.0"
description = "Publication-quality deep learning visualizations"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "numpy>=1.24",
    "scipy>=1.10",
    "scikit-learn>=1.2",
    "umap-learn>=0.5",
]

[project.optional-dependencies]
torch = ["torch>=2.0"]
transformers = ["torch>=2.0", "transformers>=4.30"]
dev = ["pytest>=7.0"]
all = ["visual-deep-learning[torch,transformers,dev]"]

[tool.setuptools.packages.find]
include = ["lib*", "loss_curve*", "tsne*", "umap_viz*", "embedding_drift*", "attention*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
